end recurring

+++ src/main/java/com/jss/osiris/modules/invoicing/model/InvoiceKeyGenerator.java
@@ -1,7 +1,6 @@
 package com.jss.osiris.modules.invoicing.model;
 
 import java.io.Serializable;
-import java.math.BigInteger;
 
 import org.hibernate.HibernateException;
 import org.hibernate.engine.spi.SharedSessionContractImplementor;
@@ -19,7 +18,7 @@ public class InvoiceKeyGenerator implements IdentifierGenerator {
         } else {
             seqName = "invoice_sequence";
         }
-        return ((BigInteger) session.createNativeQuery("SELECT nextval('" + seqName + "')").getSingleResult())
+        return ((Long) session.createNativeQuery("SELECT nextval('" + seqName + "')").getSingleResult())
                 .intValue();
     }
 }
